Move Grouper class into own file.
